+++ neoforge/src/main/java/de/cech12/bucketlib/mixin/NeoforgeItemMixin.java
@@ -2,10 +2,8 @@ package de.cech12.bucketlib.mixin;
 
 import de.cech12.bucketlib.api.BucketLibTags;
 import de.cech12.bucketlib.api.item.UniversalBucketItem;
-import de.cech12.bucketlib.item.UniversalBucketFluidHandler;
 import de.cech12.bucketlib.platform.Services;
 import de.cech12.bucketlib.util.BucketLibUtil;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -14,8 +12,6 @@ import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.material.Fluids;
 import net.neoforged.neoforge.common.CommonHooks;
-import net.neoforged.neoforge.common.capabilities.ICapabilityProvider;
-import net.neoforged.neoforge.items.wrapper.ShulkerItemStackInvWrapper;
 import net.neoforged.neoforge.common.extensions.IItemExtension;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
@@ -59,18 +55,6 @@ public interface NeoforgeItemMixin {
             }
         }
         return enchantment.category.canEnchant(stack.getItem());
-    }
-
-    /**
-     * @author Cech12
-     * @reason BucketLib class re-usability
-     */
-    @Overwrite
-    default @Nullable ICapabilityProvider initCapabilities(ItemStack stack, @Nullable CompoundTag nbt) {
-        if (this instanceof UniversalBucketItem) {
-            return new UniversalBucketFluidHandler(stack);
-        }
-        return ShulkerItemStackInvWrapper.createDefaultProvider(stack);
     }
 
 }